// src/types/order.type .ts
import { DeliveryType } from './delivery.type';
import { OrdersStatusType } from './orders-status.type';
import { PaymentType } from './payment.type ';

export type OrderType = {
  deliveryType: DeliveryType;
  firstName: string;
  lastName: string;
  fatherName?: string;
  phone: string;
  paymentType: PaymentType;
  email: string;
  street?: string;
  house?: string;
  entrance?: string;
  apartment?: string;
  comment?: string;
  items?: {
    id: string;
    name: string;
    quantity: number;
    price: number;
    total: number;
  }[];
  status?: OrdersStatusType;
  totalAmount?: number;

  statusRus?: string;
  color?: string;
};